APM跟踪器中间件独立
