+++ NewLife.CubeNC/WebMiddleware/RunTimeMiddleware.cs
@@ -1,18 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
-using NewLife.Collections;
 using NewLife.Common;
-using NewLife.Log;
-using NewLife.Model;
 using NewLife.Security;
-using NewLife.Threading;
 using XCode.DataAccessLayer;
 using XCode.Membership;
 
@@ -26,17 +20,6 @@ namespace NewLife.Cube.WebMiddleware
         /// <summary>会话提供者</summary>
         static readonly SessionProvider _sessionProvider = new SessionProvider();
 
-        /// <summary>跟踪器</summary>
-        public static ITracer Tracer { get; set; }
-
-#if DEBUG
-        static RunTimeMiddleware()
-        {
-            Tracer = DefaultTracer.Instance;
-            DAL.GlobalTracer = DefaultTracer.Instance;
-        }
-#endif
-
         /// <summary>实例化</summary>
         /// <param name="next"></param>
         public RunTimeMiddleware(RequestDelegate next) => _next = next ?? throw new ArgumentNullException(nameof(next));
@@ -46,15 +29,6 @@ namespace NewLife.Cube.WebMiddleware
         /// <returns></returns>
         public async Task Invoke(HttpContext ctx)
         {
-            // APM跟踪
-            //var span = Tracer?.NewSpan(ctx.Request.Path);
-            ISpan span = null;
-            if (Tracer != null)
-            {
-                var action = GetAction(ctx);
-                if (!action.IsNullOrEmpty()) span = Tracer.NewSpan(action);
-            }
-
             ManageProvider.UserHost = ctx.GetUserHost();
 
             // 创建Session集合
@@ -82,16 +56,9 @@ namespace NewLife.Cube.WebMiddleware
             {
                 await _next.Invoke(ctx);
             }
-            catch (Exception ex)
-            {
-                span?.SetError(ex, ctx.Request.QueryString + "");
-
-                throw;
-            }
             finally
             {
                 sw.Stop();
-                span?.Dispose();
 
                 DAL.LocalFilter = null;
                 ManageProvider.UserHost = null;
@@ -142,26 +109,6 @@ namespace NewLife.Cube.WebMiddleware
                 //Session = _sessionProvider.GetSession(ss.Id);
                 ctx.Items["Session"] = _sessionProvider.GetSession(token);
             }
-        }
-
-        /// <summary>忽略的后缀</summary>
-        public static String[] ExcludeSuffixes { get; set; } = new[] {
-            ".js", ".css", ".png", ".jpg", ".gif", ".ico",  // 脚本样式图片
-            ".woff", ".woff2", ".svg", ".ttf", ".otf", ".eot"   // 字体
-        };
-
-        private static String GetAction(HttpContext ctx)
-        {
-            var p = ctx.Request.Path + "";
-            if (p.EndsWithIgnoreCase(ExcludeSuffixes)) return null;
-
-            var ss = p.Split('/');
-            if (ss.Length == 0) return p;
-
-            // 如果最后一段是数字，则可能是参数，需要去掉
-            if ((ss.Length == 4 || ss.Length == 5) && ss[^1].ToInt() > 0) p = "/" + ss.Take(ss.Length - 1).Join("/");
-
-            return p;
         }
     }
